PR changes addressed issue #111 and #117
