+++ Middleware/src/test/java/ca/concordia/encs/citydata/BaseMvc.java
@@ -1,6 +1,5 @@
-package ca.concordia.encs.citydata.base;
+package ca.concordia.encs.citydata;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -11,7 +10,7 @@ import org.springframework.test.web.servlet.MockMvc;
  * Provides common setup and configuration for web layer testing.
  * 
  * @author Sikandar Ejaz
- * @since 08-12-2025
+ * @since 2025-08-13
  */
 
 @SpringBootTest
@@ -20,16 +19,5 @@ public abstract class BaseMvc {
 
 	@Autowired
 	protected MockMvc mockMvc;
-
-	@BeforeEach
-	public void baseSetUp() {
-		performCommonSetup();
-	}
-
-	protected void performCommonSetup() {
-	}
-
-	protected void performCommonCleanup() {
-	}
 
 }
