+++ Middleware/src/test/java/ca/concordia/encs/citydata/core/AppHealthTest.java
@@ -11,7 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.web.context.WebApplicationContext;
 
-import ca.concordia.encs.citydata.base.BaseMvc;
+import ca.concordia.encs.citydata.BaseMvc;
 import ca.concordia.encs.citydata.core.configs.AppConfig;
 
 /**
@@ -22,7 +22,7 @@ import ca.concordia.encs.citydata.core.configs.AppConfig;
  * 
  * Last Update: Removed local mockMvc instance, used from BaseMvc
  * @author Sikandar Ejaz
- * @since 12-08-2025
+ * @since 2025-08-12
  */
 
 @SpringBootTest(classes = AppConfig.class)
