+++ Middleware/src/test/java/ca/concordia/encs/citydata/core/ApplyTest.java
@@ -30,7 +30,6 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
 import ca.concordia.encs.citydata.PayloadFactory;
-import ca.concordia.encs.citydata.TestTokenGenerator;
 import ca.concordia.encs.citydata.core.configs.AppConfig;
 import ca.concordia.encs.citydata.core.utils.ReflectionUtils;
 import ca.concordia.encs.citydata.services.TokenService;
@@ -43,7 +42,7 @@ import ca.concordia.encs.citydata.services.TokenService;
  * 
  * Fixed failing tests after implementing Authentication
  * Author: Sikandar Ejaz 
- * Date: 18-07-2025
+ * Date: 2025-07-18
  */
 
 @SpringBootTest(classes = AppConfig.class)
