+++ Middleware/src/test/java/ca/concordia/encs/citydata/core/DiscoveryRoutesTest.java
@@ -3,6 +3,7 @@ package ca.concordia.encs.citydata.core;
 import static org.hamcrest.Matchers.containsString;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import org.junit.jupiter.api.Test;
@@ -10,7 +11,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.ComponentScan;
 
-import ca.concordia.encs.citydata.base.BaseMvc;
+import ca.concordia.encs.citydata.BaseMvc;
 import ca.concordia.encs.citydata.core.configs.AppConfig;
 
 /**
@@ -21,7 +22,7 @@ import ca.concordia.encs.citydata.core.configs.AppConfig;
  * 
  * Last Update: Removed local mockMvc instance, used from BaseMvc
  * @author Sikandar Ejaz
- * @since 12-08-2025
+ * @since 2025-08-12
  */
 
 @SpringBootTest(classes = AppConfig.class)
@@ -46,6 +47,31 @@ public class DiscoveryRoutesTest extends BaseMvc {
 	public void testProducersList() throws Exception {
 		mockMvc.perform(get("/producers/list")).andExpect(status().isOk())
 				.andExpect(content().string(containsString("ca.concordia.encs.citydata")));
+	}
+
+	@Test
+	void testListOperationsController() throws Exception {
+		mockMvc.perform(get("/operations/list")).andExpect(status().is2xxSuccessful())
+				.andExpect(content().string(containsString("ca.concordia.encs.citydata.operations.MergeOperation")))
+				.andExpect(content().string(containsString("targetProducerParams")))
+				.andExpect(content().string(containsString("targetProducer")));
+	}
+
+	@Test
+	void testListProducerController() throws Exception {
+		mockMvc.perform(get("/producers/list")).andExpect(status().is2xxSuccessful())
+				.andExpect(content()
+						.string(containsString("ca.concordia.encs.citydata.producers.EnergyConsumptionProducer")))
+				.andExpect(content().string(containsString("operation")))
+				.andExpect(content().string(containsString("city")));
+	}
+
+	@Test
+	void testRouteController() throws Exception {
+		mockMvc.perform(get("/routes/list")).andExpect(status().is2xxSuccessful()).andExpect(jsonPath("$").isArray())
+				.andExpect(jsonPath("$[?(@ =~ /.*\\/operations\\/list.*/)]").exists())
+				.andExpect(content().string(containsString("/operations/list")))
+				.andExpect(content().string(containsString("Method: [GET]")));
 	}
 
 }
